command debug:router fixed by asking for domain
- routes are domain-specific in Shopsys Framework but there is no current domain in console environment
- RouterDebugCommandForDomain extends the original Symfony RouterDebugCommand
- the logic of choosing a domain used in ServerRunForDomain extracted into DomainChoiceHandler for reuse
- the debug:router now asks for domain in interactive mode (or chooses the first domain by default)
