+++ packages/framework/src/Component/Console/Exception/ConsoleException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Shopsys\FrameworkBundle\Component\Console\Exception;
+
+use Throwable;
+
+interface ConsoleException extends Throwable
+{
+}
